moved detect_intervals to dedicated script
